split up covector tests
